update aws_ami filter
We came across an issue where AWS released the v1.11 AMI _before_ they
released the v1.10 AMI. So, simply choosing the latest AMI actually
selected the newest v1.10 AMI for a v1.11 cluster. If we include the k8s
version number we can ensure we're getting the latest AMI for the
cluster version.